Update in response to code feedback - Mostly small changes - Some moving of code - Made MetadataToken property across ecma reflection objects as it was an oversight I didn't do that before
