+++ src/System.Private.Reflection.Core/src/System/Reflection/Runtime/MethodInfos/RuntimeMethodInfo.cs
@@ -173,7 +173,7 @@ namespace System.Reflection.Runtime.MethodInfos
 
         public abstract override MethodInfo MakeGenericMethod(params Type[] typeArguments);
 
-        public sealed override int MetadataToken
+        public override int MetadataToken
         {
             get
             {
