+++ src/System.Private.Reflection.Core/src/System/Reflection/Runtime/TypeInfos/RuntimeTypeInfo.cs
@@ -356,7 +356,7 @@ namespace System.Reflection.Runtime.TypeInfos
             }
         }
 
-        public sealed override int MetadataToken
+        public override int MetadataToken
         {
             get
             {
